Revert "Working on serveral tasks, database creation at startup, fast admin i…"
This reverts commit 8ef85af0f252bb9c814d10333f5adb62230ebd60.
