+++ api-constantec/Autenticacion/seguridad.py
@@ -1,10 +1,10 @@
-from datetime import datetime, timedelta, timezone
-from typing import Annotated, Any, Dict, Optional
 
-from fastapi import Depends, HTTPException, status
-from fastapi.security import OAuth2PasswordBearer
-from jose import JWTError, jwt
 from passlib.context import CryptContext
+from typing import Dict, Any, Optional, Annotated
+from fastapi import Depends, HTTPException, status
+from datetime import datetime, timedelta, timezone
+from jose import JWTError, jwt
+from fastapi.security import OAuth2PasswordBearer
 
 SECRET_KEY = "your-very-secret-key-that-should-be-long-and-random"
 ALGORITHM = "HS256"
@@ -15,29 +15,21 @@ pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")
 
 oauth2_scheme = OAuth2PasswordBearer(tokenUrl="token")
 
-
 def verify_password(plain_password: str, hashed_password: str) -> bool:
     return pwd_context.verify(plain_password, hashed_password)
-
 
 def get_password_hash(password: str) -> str:
     return pwd_context.hash(password)
 
-
-def create_access_token(
-    jwt_payload: Dict[str, Any], expires_delta: Optional[timedelta] = None
-) -> str:
+def create_access_token(jwt_payload: Dict[str, Any], expires_delta: Optional[timedelta] = None) -> str:
     to_encode = jwt_payload.copy()
     if expires_delta:
         expire = datetime.now(timezone.utc) + expires_delta
     else:
-        expire = datetime.now(timezone.utc) + timedelta(
-            minutes=ACCESS_TOKEN_EXPIRE_MINUTES
-        )
+        expire = datetime.now(timezone.utc) + timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES)
     to_encode.update({"exp": expire})
     encoded_jwt = jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
     return encoded_jwt
-
 
 def decode_access_token(token: str) -> Optional[dict[str, Any]]:
     try:
@@ -45,18 +37,17 @@ def decode_access_token(token: str) -> Optional[dict[str, Any]]:
         return payload
     except JWTError:
         return None
+    
 
-
-async def get_current_user(
-    token: Annotated[str, Depends(oauth2_scheme)],
-) -> dict[str, Any]:
+async def get_current_user(token: Annotated[str, Depends(oauth2_scheme)]) -> dict[str, Any]:
     payload = decode_access_token(token)
-
+    
     if payload is None or payload.get("sub") is None:
         raise HTTPException(
             status_code=status.HTTP_401_UNAUTHORIZED,
             detail="Could not validate credentials",
             headers={"WWW-Authenticate": "Bearer"},
         )
-
+    
     return payload
+
