+++ api-constantec/Database/fake_data.py
@@ -1,0 +1,5 @@
+from Models.factories import EstudiantesFactory
+from Database.database import Base
+
+def crear_estudiantes(numero_estudiantes: int):
+    EstudiantesFactory.create_batch(5)
