+++ api-constantec/Schemas/schemas.py
@@ -1,8 +1,6 @@
-from datetime import date
-from typing import List, Optional
-
 from pydantic import BaseModel
-
+from typing import List, Optional
+from datetime import date
 
 class ConstanciaTiposSchema(BaseModel):
     id: int
@@ -10,22 +8,19 @@ class ConstanciaTiposSchema(BaseModel):
     descripcion: str
 
     class Config:
-        from_attributes = True
-
+        orm_mode = True
 
 class ConstanciaOpcionesSchema(BaseModel):
     id: int
-    tipo: ConstanciaTiposSchema
+    tipo: ConstanciaTiposSchema 
 
     class Config:
-        from_attributes = True
-
+        orm_mode = True
 
 class ConstanciaSchema(BaseModel):
     descripcion: Optional[str] = None
     otros: Optional[str] = None
     opciones: List[ConstanciaOpcionesSchema] = []
-
 
 class EstudianteSchema(BaseModel):
     no_control: str
@@ -36,15 +31,13 @@ class EstudianteSchema(BaseModel):
     municipio: str
     correo_institucional: str
 
-
 class SolicitudEstatusSchema(BaseModel):
     id: int
     tipo: str
     descripcion: str
 
     class Config:
-        from_attributes = True
-
+        orm_mode = True
 
 class SolicitudRequestSchema(BaseModel):
     fecha_solicitud: date
@@ -53,8 +46,7 @@ class SolicitudRequestSchema(BaseModel):
     constancia: ConstanciaSchema
 
     class Config:
-        from_attributes = True
-
+        orm_mode = True
 
 class SolicitudResponseSchema(BaseModel):
     fecha_entrega: Optional[date] = None
@@ -64,8 +56,7 @@ class SolicitudResponseSchema(BaseModel):
     constancia: ConstanciaSchema
 
     class Config:
-        from_attributes = True
-
+        orm_mode = True
 
 class EstudianteBase(BaseModel):
     no_control: str
@@ -78,24 +69,19 @@ class EstudianteBase(BaseModel):
     fecha_registro: date
     primer_ingreso: Optional[bool] = True
 
-
 class EstudiantesLogin(BaseModel):
     no_control: str
     contrasena: str
-
 
 class EstudiantesContrasenaUpdate(BaseModel):
     No_Control: str
     Nueva_Contrasena: str
 
-
 class EstudiantesSalida(EstudianteBase):
     class Config:
-        from_attributes = True
-
+        orm_mode = True
 
 # Constancias
-
 
 class CrearConstanciaRequest(BaseModel):
     descripcion: str
@@ -103,20 +89,16 @@ class CrearConstanciaRequest(BaseModel):
     id_estudiante: int
     constancia_opciones: list[int]
 
-
 class ConstanciaBase(BaseModel):
     Tipo: str
     Descripcion: str
     Requisitos: str
 
-
 class ConstanciaSalida(ConstanciaBase):
     class Config:
-        from_attributes = True
-
+        orm_mode = True
 
 # Solicitudes
-
 
 class SolicitudPost(BaseModel):
     No_Control: str
@@ -126,7 +108,6 @@ class SolicitudPost(BaseModel):
     Fecha_Entrega: Optional[date] = None
     ID_Trabajador: str
 
-
 class SolicitudBase(BaseModel):
     estudiantes_id: int
     solicitud_estatus_id: int
@@ -134,23 +115,18 @@ class SolicitudBase(BaseModel):
     fecha_entrega: Optional[date] = None
     # id_trabajador: str
 
-
 class SolicitudEstado(BaseModel):
     ID_Solicitud: int
-
 
 class SolicitudNuevoEstado(BaseModel):
     ID_Solicitud: str
     Nuevo_Estado: str
 
-
 class SolicitudSalida(SolicitudBase):
     class Config:
-        from_attributes = True
-
+        orm_mode = True
 
 # Historial de solicitudes
-
 
 class HistorialSolicitudBase(BaseModel):
     ID_Historial: str
@@ -159,14 +135,11 @@ class HistorialSolicitudBase(BaseModel):
     Estado_Actual: str
     Fecha_Cambio: date
 
-
 class HistorialSolicitudSalida(HistorialSolicitudBase):
     class Config:
-        from_attributes = True
-
+        orm_mode = True
 
 # Trabajadores
-
 
 class TrabajadorBase(BaseModel):
     ID_Trabajador: str
@@ -179,7 +152,6 @@ class TrabajadorBase(BaseModel):
     Correo_Institucional: str
     Fecha_Inicio: date
 
-
 class TrabajadorSalida(TrabajadorBase):
     class Config:
-        from_attributes = True
+        orm_mode = True
