+++ api-constantec/main.py
@@ -1,0 +1,38 @@
+from fastapi import FastAPI, HTTPException
+from fastapi.responses import FileResponse
+from fastapi.security import HTTPBearer, HTTPAuthorizationCredentials
+from fastapi.staticfiles import StaticFiles
+import logging
+from fastapi.middleware.cors import CORSMiddleware
+from Routers import estudiantes, constancias, solicitudes, login
+
+logging.basicConfig(
+    level=logging.DEBUG,
+    format="%(asctime)s [%(levelname)s] %(name)s: %(message)s",
+)
+
+logger = logging.getLogger(__name__)
+
+app = FastAPI()
+
+app.add_middleware(
+    CORSMiddleware,
+    allow_origins=["http://localhost:5173"],  # Vite dev server default
+    allow_credentials=True,
+    allow_methods=["*"],
+    allow_headers=["*"],
+)
+
+app.mount("/assets", StaticFiles(directory="web-app/assets"), name="assets")
+
+app.include_router(login.router, prefix="/v1/login", tags=["Login"])
+app.include_router(estudiantes.router, prefix="/v1/estudiantes", tags=["Estudiantes"])
+app.include_router(constancias.router, prefix="/v1/constancias", tags=["Constancias"])
+app.include_router(solicitudes.router, prefix="/v1/solicitudes", tags=["Solicitudes"])
+
+
+@app.get("/{full_path:path}")
+def iniciando_sesion(full_path: str):
+    if full_path.startswith("v1"):
+        raise HTTPException(status_code=404, detail="Not Found")
+    return FileResponse("web-app/index.html")
