+++ api-constantec/requirements_dev.txt
@@ -2,18 +2,14 @@ fastapi
 uvicorn[standard]
 black
 isort
-pylint
 sqlalchemy
 Schemas
 pydantic
 typing
 datetime
 pyodbc
-aioodbc
 passlib[bcrypt]
-bcrypt==3.2.0
-python-jose
+bcrypt<4.0
+python-jose[cryptography]
 factory_boy
 ipython
-asyncpg
-sqladmin
