--- constantec/admin/__init__.py
@@ -1,4 +1,0 @@
-from .estudiantes import EstudiantesAdmin
-from .admin_users import UserAdmin
-
-__all__ = ['EstudiantesAdmin', 'UserAdmin']
