+++ docker-compose-dev.yml
@@ -5,7 +5,7 @@ services:
     ports:
       - "8000:8000"
     volumes:
-      - ./constantec:/app
+      - ./api-constantec:/app
     environment:
       - DB_HOST=sqlserver
       - DB_PORT=1433
@@ -13,8 +13,6 @@ services:
       - DB_PASSWORD=YourStrong!Passw0rd
       - DB_NAME=constantec
       - PYTHONUNBUFFERED=1
-    depends_on:
-      - sqlserver
   
   sqlserver:
     image: mcr.microsoft.com/mssql/server:2022-latest
